requirements.txt
playwright==1.45.0
pytesseract==0.3.10
Pillow==10.3.0
pandas==2.2.2
requests==2.31.0
opencv-python==4.10.0.84
numpy==1.26.4
